+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * success toasts added;
     * cards loading behavior added;
     * photosets appearance improved;
+    * the 'Trash' functionality added for photos (remove photos into Trash first; photos can be then restored from Trash or removed permanently);
 * `Facade` - homepage updated;
 
 ## **1.0.0** - *2019-10-26*